Move getTokenWithPopup tests to Auth0Client.test.ts
