+++ __tests__/index.test.ts
@@ -1156,112 +1156,6 @@ describe('Auth0', () => {
     });
   });
 
-  describe('getTokenWithPopup()', () => {
-    const localSetup = async (clientOptions?: Partial<Auth0ClientOptions>) => {
-      const result = await setup(clientOptions);
-      result.auth0.loginWithPopup = jest.fn();
-      result.cache.get.mockReturnValue({ access_token: TEST_ACCESS_TOKEN });
-      return result;
-    };
-
-    it('calls `loginWithPopup` with the correct default options', async () => {
-      const { auth0 } = await localSetup();
-
-      await auth0.getTokenWithPopup();
-
-      expect(auth0.loginWithPopup).toHaveBeenCalledWith(
-        {
-          audience: undefined,
-          scope: TEST_SCOPES
-        },
-        DEFAULT_POPUP_CONFIG_OPTIONS
-      );
-    });
-
-    it('respects customized default scopes', async () => {
-      const { auth0 } = await localSetup({
-        advancedOptions: {
-          defaultScope: 'email'
-        }
-      });
-
-      await auth0.getTokenWithPopup();
-
-      expect(auth0.loginWithPopup).toHaveBeenCalledWith(
-        {
-          audience: undefined,
-          scope: 'openid email'
-        },
-        DEFAULT_POPUP_CONFIG_OPTIONS
-      );
-    });
-
-    it('calls `loginWithPopup` with the correct custom options', async () => {
-      const { auth0 } = await localSetup();
-      const loginOptions = {
-        audience: 'other-audience',
-        scope: 'other-scope'
-      };
-
-      const configOptions = { timeoutInSeconds: 1 };
-
-      await auth0.getTokenWithPopup(loginOptions, configOptions);
-
-      expect(auth0.loginWithPopup).toHaveBeenCalledWith(
-        {
-          audience: 'other-audience',
-          scope: `${TEST_SCOPES} other-scope`
-        },
-        configOptions
-      );
-    });
-
-    it('calls `cache.get` with the correct options', async () => {
-      const { auth0, cache } = await localSetup();
-
-      await auth0.getTokenWithPopup();
-
-      expect(cache.get).toHaveBeenCalledWith({
-        audience: 'default',
-        scope: TEST_SCOPES,
-        client_id: TEST_CLIENT_ID
-      });
-    });
-
-    it('returns cached access_token', async () => {
-      const { auth0 } = await localSetup();
-
-      const token = await auth0.getTokenWithPopup();
-      expect(token).toBe(TEST_ACCESS_TOKEN);
-    });
-
-    it('accepts empty options and config', async () => {
-      const { auth0 } = await localSetup({ audience: 'foo' });
-
-      await auth0.getTokenWithPopup();
-      expect(auth0.loginWithPopup).toHaveBeenCalledWith(
-        {
-          audience: 'foo',
-          scope: 'openid profile email'
-        },
-        { timeoutInSeconds: 60 }
-      );
-    });
-
-    it('accepts partial options and config', async () => {
-      const { auth0 } = await localSetup({ audience: 'foo' });
-
-      await auth0.getTokenWithPopup({ scope: 'bar' }, { popup: 'baz' });
-      expect(auth0.loginWithPopup).toHaveBeenCalledWith(
-        {
-          audience: 'foo',
-          scope: 'openid profile email bar'
-        },
-        { timeoutInSeconds: 60, popup: 'baz' }
-      );
-    });
-  });
-
   describe('buildLogoutUrl()', () => {
     it('creates correct query params with empty options', async () => {
       const { auth0, utils } = await setup();
